Allow upload photo on the user model
Allow upload photo on the user model
