--- photos/photo.model.js
@@ -1,18 +1,0 @@
-const mongoose = require('mongoose');
-const Schema = mongoose.Schema;
-
-const PhotoSchema = new Schema({
-  userID: { type: String, required: true },
-  photo: { type: String},
-});
-
-PhotoSchema.set('toJSON', {
-  virtuals: true,
-  versionKey: false,
-  transform: function (doc, ret) {
-      delete ret._id;
-      delete ret.hash;
-  }
-});
-
-module.exports = mongoose.model('Photo', PhotoSchema);
